Made a new recipe
Added a new recipe on saucy bean baked eggs. the page is structured
like the carbonara page, as in it starts with the headder line explaning
what recipe it is and thereafter a pic of said food. It contiues with an
ingrediens list then an method list. Thereafter I added into the index.html file.